% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail.r
\name{gail}
\alias{gail}
\title{Geo-Assignment of Irregular Locations}
\usage{
gail(sp_units, cases, suid, num_cases, max_dist, group, labels,
  RAP = gail_rap, seed = NULL, convert = FALSE, ...)
}
\arguments{
\item{sp_units}{Data frame of regular spatial units, must have columns named \code{longitude} and \code{latitude}.}

\item{cases}{Data frame containing the the cases, see details.}

\item{suid}{Column name in both \code{sp_units} and \code{cases} which contains the spatial unit identification / name.}

\item{num_cases}{Column name from \code{cases} containing the number of cases, see details.}

\item{max_dist}{The maximum distance at which two locations can be considered neighbors.}

\item{group}{Name of variable in \code{sp_units} to denote regular and irregular spatial units.}

\item{labels}{Vector of values for the regular and irregular units (the values from \code{group}), with the first.
element being the regular units, and the second element being the irregular units.}

\item{RAP}{Function to be used for calculating assignment probabilities, see details.}

\item{seed}{If given, sets the seed for the RNG.}

\item{convert}{Logical, convert returned object to \code{Spatial*} type objects (from \strong{sp}).}

\item{...}{Space for additional arguments (e.g., for \code{RAP}).}
}
\value{
Returns a list containing DESCRIBE THE RESULTS
}
\description{
DESCRIPTION OF FUNCTION
}
\details{
\itemize{
\item \code{cases} : The dataset of cases should be aggregated and have columns for the spatial unit name
(given by \code{suid}) and the number of cases. If the column given by \code{num_cases} does not exist in \code{cases},
then \code{cases} will be aggregated to create the column.
\item \code{RAP} : The argument controls how irregular spatial units are allocated to regular spatial units.
There are several formats this can take. If \code{RAP} is a function, it will be used. Otherwise \code{gail_rap}
is used. By default the internal function \code{gail_rap} is used with arguments \code{method} and \code{index_var}
(see documentation of \link{gail_rap}) determined by:
\itemize{
\item If \code{RAP} is a column name from \code{sp_units} then \code{method="index"} and \code{index_var = RAP} .
\item If \code{RAP="icd"} then \code{method="icd"} and \code{index_var = NULL} .
\item Otherwise, \code{method="equal"} and \code{index_var = NULL} .
}
}

The user may also specify only \code{method} and/or \code{index_val} to control the behavior of \code{gail_rap} appropriately.
\itemize{
\item \code{labels} : The general (but not default) format of this is \code{labels=c("Standard", "Irregular")}. If \code{labels}
is missing, \code{gail} finds the unique values of \code{group} and takes the first element as the regular spatial
units, and the second element as the irregular spatial units.
}
}
\note{
MAYBE SOME COMMENTS.
}
