% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_mc.r
\name{gail_mc}
\alias{gail_mc}
\title{Monte Carlo Estimation of Variability from GAIL}
\usage{
gail_mc(data)
}
\arguments{
\item{data}{the data ... replace me}
}
\value{
Returns a list containing DESCRIBE THE RESULTS
}
\description{
Monte Carlo Estimation of Variability from GAIL
}
