% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{gail_nn}
\alias{gail_nn}
\title{Number of Neighbors}
\usage{
gail_nn(val, df1, df2, max_dist, suid, ...)
}
\arguments{
\item{val}{Name of spatial unit from df2.}

\item{df1}{Data frame of regular spatial units, must have columns \code{latitude} and \code{longitude}.}

\item{df2}{Data frame of irregular spatial units, must have columns named \code{latitude} and \code{longitude}.}

\item{max_dist}{The maximum distance at which two locations can be considered neighbors.}

\item{suid}{The column name which contains the ID of the spatial unit.}

\item{...}{Space for additional arguments, passed to \link[sf:st_distance]{sf::st_distance}.}
}
\value{
Returns the number of spatial units from \code{df1} that are within
distance \code{max_dist} of the unit \code{val} from \code{df2}.
}
\description{
Calculates the the number of neighbors for a spatial unit.
}
\details{
This is an internal function designed to be used within \link{gail}
as part of a purrr type operation.
}
\seealso{
\link{gail}
}
