% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{gail_rap}
\alias{gail_rap}
\title{Region Assignment Probability}
\usage{
gail_rap(rUnits, iUnits, rUclose, rUdist, max_dist, method, index_val)
}
\arguments{
\item{rUnits}{Data frame of regular spatial units.}

\item{iUnits}{Data frame of irregular spatial unit being allocated.}

\item{rUclose}{The set of regular units that are considered close (within \code{max_dist})}

\item{rUdist}{Vector of distances of \code{rUclose} to \code{iUnits}}

\item{max_dist}{The maximum distance at which two locations can be considered neighbors.}

\item{method}{The method of allocation, see details.}

\item{index_val}{If \code{method="index"}, the name of the variable (in \code{rUnits}) to be used as an index variable.}
}
\value{
Returns a set of weights which has length equal to the number of spatial units in \code{rUclose}.
}
\description{
Function to calculate the (relative) probability of assigning an irregular unit to a
neighboring regular unit.
}
\details{
\itemize{
\item \code{method = "index"} : the variable given by \code{index_val} (in \code{rUnits}) is used in the form \code{X / sum(X)}.
\item \code{method = "equal"} : all neighboring spatial units (in \code{rUclose}) are given equal probability.
\item \code{method = "icd"}   : inverse centroid distance, the closer the centroid of a regular spatial unit,
the higher the probability. If \code{D} is the distance to the centroid, the probabilities are given by:
\code{(1/D) / sum(1/D) = 1 / ( D * sum(1/D) )}. Similar to the index method, but uses a computed variable.
}
}
\note{
This is an internal function designed to be used within \link{gail}. This function
provides several default options for the \code{RAP} argument to \code{gail}. Users can alternatively
supply a custom function as the argument for the \code{RAP} argument in \code{gail}, provided it accepts
the same arguments and the returns is the same type and dimension.

The returned value is passed to \code{sample} as the argument \code{prob}.
}
\seealso{
\link{gail}
}
