% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_shp.r
\name{gail_shp}
\alias{gail_shp}
\title{Geo-Assignment of Irregular Locations to Shapes}
\usage{
gail_shp(sp_units, cases, suid, num_cases, group, label, convert = FALSE,
  ...)
}
\arguments{
\item{sp_units}{Dataset of regular spatial units, see details.}

\item{cases}{Dataset containing the the cases, see details.}

\item{suid}{Column name in both \code{sp_units} and \code{cases} which contains the spatial unit identification / name.}

\item{num_cases}{Column name from \code{cases} containing the number of cases, see details.}

\item{group}{Optional name of variable in \code{cases} to use for filtering. See details.}

\item{label}{Optional value of \code{group} by which to filter. See details.}

\item{convert}{Logical, convert returned object to \code{Spatial*} type objects (from \strong{sp}).}

\item{...}{Space for additional arguments.}
}
\value{
Returns a copy of \code{sp_units} after having added a column named \code{num_cases}, which contains
the number of cases which fall within the boundry of each spatial unit.
}
\description{
Allocates the number of cases/events from irregular spatial units onto to regular set of spatial units.
The regular spatial units are assumed to have POLYGON geometry, while the irregular spatial units are
assumed to have POINT geometry.

Note that this function is largely a wrapper for using \code{sf::st_contains} with basic
aggregation before and after. It is provided primarily for (relative) consistency with \link{gail}.
}
\details{
\itemize{
\item \code{sp_units} should be an object of class \code{sf} (from the \strong{sf} package). The
argument \code{convert} can be used to convert the results into a \code{Spatial*} object (from
the \strong{sp} package).
\item \code{cases} contains the units (and associated number of cases) to be allocated.
This should be an object of class \code{sf}. If not, it must have coordinates \code{latitude}
and \code{longitude} and will be converted to an \code{sf} object.
The dataset of cases should be aggregated and have a column for the number
of cases. If the column given by \code{num_cases} does not exist in \code{cases},
then \code{cases} will be aggregated by \code{suid} to create the column.
\item \code{group} and \code{label} : If provided, these arguments are used to filter \code{cases}
to rows where \code{group == label}.
}
}
\seealso{
\link{gail}
}
