% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_sim.r
\name{gail_sim_index}
\alias{gail_sim_index}
\title{Generate a Simulated Similarity Index}
\usage{
gail_sim_index(units_sp, tau, phi, seed = NULL)
}
\arguments{
\item{units_sp}{Set of spatial units}

\item{tau}{Nugget variance parameter, see details.}

\item{phi}{Spatial dependence parameter, see details.}

\item{seed}{If given, sets the seed for the RNG.}
}
\description{
Function for the simulation framework in GAIL.
Creates a spatially dependent similarity index across the regions.
}
\details{
Generates a spatially-dependent similarity index \emph{y} by first simulating
from a multivariate normal distribution:

$Y \sim \mbox{MVN}( 0, \tau^{2}( I - \phi H))$

where H is the neighborhood matrix, and I is the identity matrix. Then
the index is converted to a \link{0, 15} uniform random variable.
}
