% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_sim.r
\name{gail_sim_rate}
\alias{gail_sim_rate}
\title{Sets Incidence Rates for Simulated Regions}
\usage{
gail_sim_rate(units_reg, rate_base = c(0.03, 0.07), rate_spec = NULL,
  seed = NULL)
}
\arguments{
\item{units_reg}{Set of regular spatial units}

\item{rate_base}{Vector of length 2 giving upper and lower bounds for uniform distribution.
The base rate is randomly allocated between these two values.
Default is \code{c(0.03, 0.07)}.}

\item{rate_spec}{A \code{data.frame} describing coordinates and change to base rate. See details.}

\item{seed}{If given, sets the seed for the RNG.}

\item{...}{Space for additional arguments (e.g., for \code{fields::cover.design}).}
}
\value{
Returns a vector of length \code{nrow(units_reg)} which contains
}
\description{
Function for the simulation framework in GAIL.
Create the underlying rate of cases for each regular spatial unit which can follow a
user-specified pattern. This can be used to generate the rate of cases, as well as the
rate of individuals being in the irregular spatial unit.

For creating simulated data, this function can be bypassed if the user creates variables
named 'case_rate' and 'rural_rate' in the set of regular spatial units.
}
\details{
Each row of \code{rate_spec} creates a 'hotspot' (or 'coldspot') in terms of the incidence
rate of cases. That is: areas in the spatial domain which have an increased or decreased rate. This
should be a \code{data.frame} (or comprable object) with columns: \code{mx}, \code{my}, \code{ax}, \code{ay}, \code{efc}. The
hotspot is centered at the point (\code{mx}, \code{my}), while \code{ax} and \code{ay} control the size in the x and y
directions, respectively, with larger values corresponding to larger range in that dimension. The \code{efc}
value is the effect size, which acts as a multiplier for the base rate of indidence.

The incidence rate is generated by first drawing a base rate for each spatial unit
from a uniform distribution with bounds given by \code{rate_base}.

Then n=2000 points are drawn uniformly across the spatial domain (100x100 square). These points are
given a weight on the interval \code{[0, 1]}, which decreases from 1 at the center of the hotspot down to 0.
The weight is multiplied by the hotspot effect size (\code{efc}), and shifted so that each of the 2000
points have a value on the interval \code{[1, efc]}. The mean effect is taken for all individuals contained
within a spatial region, and that mean is used as a multiplier for the base rate of that region.
}
\examples{
\dontrun{
 ## Generate Regions
 loca_reg <- gail_gen_regions( npoints=40, type="regular", nedge=10, suid="reg" )
 
 ## Generate incidence rate
 rate_spec <- data.frame(
   mx  = c(25, 60), 
   my  = c(25, 80), 
   ax  = c(10, 40), 
   ay  = c(25, 20),
   efc = c( 0.15 , 4.0 )
 )
 loca_reg[["case_rate"]] <- gail_sim_rate( loca_reg,  rate_base=c(0.03,0.07), 
                                            rate_spec=rate_spec, seed=42 )
 ggplot( loca_reg ) +
   geom_sf( aes(fill=case_rate) )
 
 ## Generate rate of being in irregular locations
 irr_spec <- data.frame(
   mx  = c(85, 20, 25, 60), 
   my  = c(15, 80, 25, 80), 
   ax  = c(20, 10, 10, 40), 
   ay  = c(20, 20, 25, 20),
   efc = c(4.0, 4.0, 0.15 , 0.15 )
 )
  
 loca_reg[["rural_rate"]] <- gail_sim_rate( loca_reg,  rate_base=c(0.03,0.07), 
                                            rate_spec=rate_spec, seed=42 )
    
 ggplot( loca_reg ) +
   geom_sf( aes(fill=rural_rate) )
    
}


}
\seealso{
\link{gail_sim_regions}, \link{gail_sim_pop}, \link{gail_sim_index}, \link{gail_sim_assign}
}
