% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_sim.r
\name{gail_sim_regions}
\alias{gail_sim_regions}
\title{Generate Artificial Polygon Regions}
\usage{
gail_sim_regions(npoints, type = "irregular", nedge = 5, seed = NULL,
  suid = NULL, ...)
}
\arguments{
\item{npoints}{Number of internal points to select}

\item{type}{Type of spatial units to generate, accepts "regular" and "irregular". See details.}

\item{nedge}{Data frame containing the the cases, see details.}

\item{seed}{If given, sets the seed for the RNG.}

\item{suid}{Optional spatial unit ID, used as prefix for generated spatial units.}

\item{...}{Space for additional arguments (e.g., for \code{fields::cover.design}).}
}
\value{
Returns an object of class \code{sf} with POLYGON geometry. This represents a set of areal
}
\description{
Function for the simulation framework in GAIL.
Create a set of spatial polygon regions.
For a more comprehensive description of how to use the simulation framework,
see the vignette "GAIL Simulation."
}
\details{
The \code{type} argument works as follows:
\itemize{
\item \code{type="regular"} produces a regular grid of units with \code{nedge} units along each dimension.
\item \code{type="irregular"} randomly draws 1000 points uniformly across a 100x100 region (excluding a band of
width 5 around the edge). Then uses \code{fields::cover.design} to select \code{npoints} of them.
Creates regions using Voronoi tesselation via \code{deldir::deldir} and \code{deldir::tile.list}.
}
}
\examples{
\dontrun{
 loca_reg <- gail_gen_regions( npoints=40, type="regular"  , nedge=10, suid="reg" )
 loca_irr <- gail_gen_regions( npoints=40, type="irregular", nedge=6 , seed=42 , P=-20, Q=20 )
  
 ggplot( loca_reg ) + 
   geom_sf( aes(fill=region) )
 ggplot( loca_irr ) +
   geom_sf( aes(fill=region) )
}


}
\seealso{
\link{gail_sim_rate}, \link{gail_sim_pop}, \link{gail_sim_index}, \link{gail_sim_assign}
}
