% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Package.r
\docType{package}
\name{GAIL-package}
\alias{GAIL}
\alias{GAIL-package}
\title{Geo-Assignment of Irregular Locations}
\description{
A package for handling allocating a set of irregular spatial units (e.g. P.O. Box ZIP codes)
onto a regular set of spatial units (e.g. Standard ZIP codes, Census Tract, County, etc).
}
\details{
SOME DETAILS OF GAIL
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/jelsema/GAIL/issues}
}

}
\author{
\strong{Maintainer}: Casey M. Jelsema \email{jelsema.casey@gmail.com}

Authors:
\itemize{
  \item Hendricks. Brian
  \item Dick. Taron
}

}
