% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail.r
\name{gail}
\alias{gail}
\title{Geo-Assignment of Irregular Locations}
\usage{
gail(sp_units, cases, suid, num_cases, max_dist, RAP = gail_rap,
  seed = NULL, unit_value = "m", convert = FALSE, ...)
}
\arguments{
\item{sp_units}{Data frame of regular spatial units, must have columns named \code{longitude} and \code{latitude}.}

\item{cases}{Data frame containing the the cases, see details.}

\item{suid}{Column name in both \code{cases} which contains the spatial unit identification / name.}

\item{num_cases}{Column name from \code{cases} containing the number of cases, see details.}

\item{max_dist}{The maximum distance at which two locations can be considered neighbors.}

\item{RAP}{Function to be used for calculating assignment probabilities, see details.}

\item{seed}{If given, sets the seed for the RNG.}

\item{unit_value}{The units of distance, default to meters. See \link[units:set_units]{units::set_units}.}

\item{convert}{Logical, convert returned object to \code{Spatial*} type objects (from \strong{sp}).}

\item{...}{Space for additional arguments (e.g., for \code{RAP}).}
}
\value{
Returns a list containing DESCRIBE THE RESULTS
}
\description{
Stochastic allocation of irregular spatial units to nearby regular spatial units. Both sets of
of spatial units are assummed to be represented by their centroid (hence they should be POINT objects,
and not POLYGON objects). Several methods are available for allocating irregular units, including equal
probability, inverse centroid distance, and use of an index variable. Alternatively the user may
specify a custom function to determine allocation probabilities.
}
\details{
\itemize{
\item \code{sp_units} is is set of spatial units that are the "destination", the cases will be aggregated to
this set of units. It should contain only unique spatial units (no duplicates), and should not have
any spatial units on a different scale. It must either be an object of class \code{sf} or have variables
\code{longitude} and \code{latitude} to be convertable into an object of class \code{sf}.
\item \code{cases} contains the units (and associated number of cases) to be allocated.
This should be an object of class \code{sf}. If not, it must have coordinates \code{latitude}
and \code{longitude} and will be converted to an \code{sf} object.
The dataset of cases should be aggregated and have a column for the number
of cases. If the column given by \code{num_cases} does not exist in \code{cases},
then \code{cases} will be aggregated by \code{suid} to create the column.
\item If \code{sp_units} also has a column as given by \code{suid}, this will be assumed to be a common ID between
\code{cases} and \code{sp_units}. Exact matches by \code{suid} will be allocated deterministically. Any remaining
cases will be allocated stochastically.
\item \code{RAP} : The argument controls how irregular spatial units are allocated to regular spatial units.
There are several formats this can take. If \code{RAP} is a function, it will be used. Otherwise \code{gail_rap}
is used. By default the internal function \code{gail_rap} is used with arguments \code{method} and \code{index_var}
(see documentation of \link{gail_rap}) determined by:
\itemize{
\item If \code{RAP} is a column name from \code{sp_units} then \code{method="index"} and \code{index_var = RAP} .
\item If \code{RAP="icd"} then \code{method="icd"} and \code{index_var = NULL} .
\item Otherwise, \code{method="equal"} and \code{index_var = NULL} .
}
}

The user may also specify only \code{method} and/or \code{index_val} to control the behavior of \code{gail_rap} appropriately.
}
