% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_shp.r
\name{gail_shp}
\alias{gail_shp}
\title{Geo-Assignment of Irregular Locations to Shapes}
\usage{
gail_shp(sp_units, cases, num_cases, group, label, convert = FALSE, ...)
}
\arguments{
\item{sp_units}{Data frame of regular spatial units, see details.}

\item{cases}{Data frame containing the the cases, see details.}

\item{num_cases}{Column name from \code{cases} containing the number of cases, see details.}

\item{group}{Optional name of variable in \code{cases} to use for filtering. See details.}

\item{label}{Optional value of \code{group} by which to filter. See details.}

\item{convert}{Logical, convert returned object to \code{Spatial*} type objects (from \strong{sp}).}

\item{...}{Space for additional arguments.}
}
\value{
Returns a list containing DESCRIBE THE RESULTS
}
\description{
DESCRIPTION OF FUNCTION
}
\details{
\itemize{
\item \code{cases} : The dataset of cases should be aggregated and have a column for the number
of cases. If the column given by \code{num_cases} does not exist in \code{cases},
then \code{cases} will be aggregated to create the column.
\item \code{group} and \code{label} : If provided, these arguments are used to filter \code{cases}
to rows where \code{group == label}.
}
}
\note{
MAYBE SOME COMMENTS.
}
\seealso{
\link{gail}
}
