% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_sim.r
\name{gail_sim_assign}
\alias{gail_sim_assign}
\title{Allocate Simulated Population}
\usage{
gail_sim_assign(units_reg, units_irr, loca_pop, seed = NULL)
}
\arguments{
\item{units_reg}{Set of regular spatial units}

\item{units_irr}{Set of irregular spatial units}

\item{loca_pop}{True locations of the population}

\item{seed}{If given, sets the seed for the RNG.}
}
\description{
Function for the simulation framework in GAIL.
Determined the spatial unit (both regular and irregular) into which a population individual
falls, and randomly samples whether an individual is case vs non-case and whether an individual
is in regular or non-regular spatial unit.
}
\details{
The argument \code{units_reg} must have columns \code{case_rate} and \code{rural_rate}. These will
ordinarily be generated by other simulation functions in the process of simulating
the data.

This function assigns case/non and regular/irregular by:
\enumerate{
\item Assigning entire population to the regular spatial units
\item For each regular spatial unit, independently sampling based on case rate (to generate cases)
and rural rate (to generate individuals reporting the irregular spatial unit).
}
}
