% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_sim.r
\name{gail_sim_pop}
\alias{gail_sim_pop}
\title{Simulate Cases for GAIL}
\usage{
gail_sim_pop(units_reg, method = "uniform", npop = 1e+05,
  beta_setup = NULL, seed = NULL)
}
\arguments{
\item{units_reg}{Set of regular spatial units (cases get allocated to this set).}

\item{method}{Method of simulating population: 'uniform' or 'beta'. See details.}

\item{npop}{Size of population to simulate.}

\item{beta_setup}{If \code{method='beta'}, a \code{data.frame} describing how to distribute the locations. See details.}

\item{seed}{If given, sets the seed for the RNG.}

\item{...}{Space for additional arguments (e.g., for \code{fields::cover.design}).}
}
\description{
Create a set of simulated cases
}
\details{
For \code{method='uniform'} points are simulated uniformly across the 100x100 spatial domain.
For \code{method='irregular'} then three additional parameters are necessary: A list of centers,
and a list of values for alpha and beta. For each center, points are simulated from a beta
distribution. If alpha and beta are not provided for each center, the default values will
set the mean of the distribution of points to be the center, and the
Describe data.frame input.
}
