% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gail_sim.r
\name{gail_sim_rate}
\alias{gail_sim_rate}
\title{Sets Incidence Rates for Simulated Regions}
\usage{
gail_sim_rate(units_reg, rate_base = c(0.03, 0.07), rate_spec = NULL,
  seed = NULL)
}
\arguments{
\item{units_reg}{Set of regular spatial units}

\item{rate_base}{Vector of length 2 giving upper and lower bounds for uniform distribution.
The base rate is randomly allocated between these two values.
Default is \code{c(0.03, 0.07)}.}

\item{rate_spec}{A \code{data.frame} describing coordinates and change to base rate. See details.}

\item{seed}{If given, sets the seed for the RNG.}

\item{...}{Space for additional arguments (e.g., for \code{fields::cover.design}).}
}
\description{
Function for the simulation framework in GAIL.
Create the underlying rate of cases for each regular spatial unit. This function can be bypassed
if the user creates a variable names 'true_rate' in the set of regular spatial units.
}
\details{
Describe the data.frame input.
}
